# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Middle",
        "contact": {},
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://middle-testing.restart-technology.com/api",
            "description": " {{DEVELOPMENT SERVER}}",
            "variables": {}
        },
        {
            "url": "http://middle.restart-technology.com/api",
            "description": " {{Live SERVER}}",
            "variables": {}
        },
        {
            "url": "https://7ekayh-games.com/api",
            "description": " {{Live SERVER}}",
            "variables": {}
        },
        {
            "url": "http://localhost:8080/restart/middle/public/api",
            "description": " {{FADY LOCALHOST SERVER}}",
            "variables": {}
        },
        {
            "url": "http://localhost/projects/middle-backend/public/api",
            "description": " {{LOCALHOST SERVER}}",
            "variables": {}
        },
        {
            "url": "http://localhost:8080/projects/middle-backend/public/api",
            "description": " {{LOCALHOST SERVER}}",
            "variables": {}
        },
        {
            "url": "http://localhost/middle-backend/public/api",
            "description": " {{LOCALHOST SERVER}}",
            "variables": {}
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": " {{Ready LOCALHOST SERVER}}",
            "variables": {}
        }


    ],
    "paths": {
...
        "/store/gifts/get": {
            "get": {
                "tags": [
                    "Gift"
                ],
                "summary": "Get My Gifts",
                "operationId": "GetMyGifts",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "headers": {}
                    }
                },
                "security": [{
                    "httpBearer": []
                }],
                "deprecated": false
            }
        },
        "/store/gifts/send": {
            "post": {
                "tags": [
                    "Store"
                ],
                "summary": "Send Gift",
                "operationId": "SendGift",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "encoding": {},
                            "schema": {
                                "required": [
                                    "room_id",
                                    "gift_id",
                                    "to_user_id"
                                ],
                                "type": "object",
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 3
                                    },
                                    "gift_id": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 3
                                    },
                                    "to_user_id": {
                                        "type": "integer",
                                        "format": "int32",
                                        "example": 3
                                    }
                                }
                            }
                        }
                    },
                    "required": false,
                    "x-send-file-in-body": false
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "headers": {}
                    }
                },
                "security": [{
                    "httpBearer": []
                }],
                "deprecated": false
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "httpBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}